requirements.txt
addict==2.4.0
aliyun-python-sdk-core==2.16.0
aliyun-python-sdk-kms==2.16.5
attrs==25.3.0
Brotli @ file:///C:/b/abs_3d36mno480/croot/brotli-split_1714483178642/work
certifi @ file:///C:/b/abs_1fw_exq1si/croot/certifi_1725551736618/work/certifi
cffi==1.17.1
charset-normalizer @ file:///croot/charset-normalizer_1721748349566/work
chumpy==0.70
click==8.1.8
colorama @ file:///C:/b/abs_a9ozq0l032/croot/colorama_1672387194846/work
contourpy==1.1.1
coverage==7.6.1
crcmod==1.7
cryptography==45.0.3
cycler==0.12.1
Cython==3.1.2
exceptiongroup==1.3.0
filelock==3.14.0
flake8==7.1.2
fonttools==4.57.0
fsspec==2025.3.0
fvcore @ file:///home/conda/feedstock_root/build_artifacts/fvcore_1671623667463/work
idna @ file:///C:/b/abs_aad84bnnw5/croot/idna_1714398896795/work
importlib_metadata==8.5.0
importlib_resources==6.4.5
iniconfig==2.1.0
interrogate==1.7.0
iopath @ file:///home/conda/feedstock_root/build_artifacts/iopath_1723846728822/work
isort==4.3.21
Jinja2==3.1.4
jmespath==0.10.0
json-tricks==3.17.3
kiwisolver==1.4.7
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.7.5
mccabe==0.7.0
mdurl==0.1.2
mkl-fft @ file:///C:/b/abs_19i1y8ykas/croot/mkl_fft_1695058226480/work
mkl-random @ file:///C:/b/abs_edwkj1_o69/croot/mkl_random_1695059866750/work
mkl-service==2.4.0
mmcv==2.1.0
mmdet==3.2.0
mmengine==0.10.7
-e git+https://github.com/open-mmlab/mmpose.git@71ec36ebd63c475ab589afc817868e749a61491f#egg=mmpose
model-index==0.1.11
mpmath==1.3.0
munkres==1.1.4
networkx==3.0
numpy==1.24.1
olefile @ file:///home/conda/feedstock_root/build_artifacts/olefile_1701735466804/work
opencv-python==4.11.0.86
opendatalab==0.0.10
openmim==0.3.9
openxlab==0.1.2
ordered-set==4.1.0
oss2==2.17.0
packaging==24.2
pandas==2.0.3
parameterized==0.9.0
Pillow==9.3.0
platformdirs==4.3.6
pluggy==1.5.0
portalocker @ file:///D:/bld/portalocker_1720928680820/work
py==1.11.0
pycocotools==2.0.7
pycodestyle==2.12.1
pycparser==2.22
pycryptodome==3.23.0
pyflakes==3.2.0
Pygments==2.19.1
pyparsing==3.1.4
PySocks @ file:///C:/ci/pysocks_1605287845585/work
pytest==8.3.5
pytest-runner==6.0.1
python-dateutil==2.9.0.post0
pytz==2023.4
pywin32==306
PyYAML @ file:///C:/b/abs_14xkfs39bx/croot/pyyaml_1728657968772/work
regex==2024.11.6
requests==2.28.2
rich==13.4.2
scipy==1.10.1
shapely==2.0.7
six==1.17.0
sympy==1.13.3
tabulate @ file:///C:/b/abs_21rf8iibnh/croot/tabulate_1701354830521/work
termcolor @ file:///home/conda/feedstock_root/build_artifacts/termcolor_1704357939450/work
terminaltables==3.1.10
tomli==2.2.1
torch==2.4.1+cu118
torchaudio==2.4.1+cu118
torchvision==0.19.1+cu118
tqdm==4.65.2
typing_extensions @ file:///C:/b/abs_0as9mdbkfl/croot/typing_extensions_1715268906610/work
tzdata==2025.2
urllib3==1.26.20
win-inet-pton @ file:///C:/ci/win_inet_pton_1605306167264/work
xdoctest==1.2.0
xtcocotools==1.14.3
yacs @ file:///home/conda/feedstock_root/build_artifacts/yacs_1645705974477/work
yapf==0.43.0
zipp==3.20.2
